requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
influxdb-client==1.37.0
pydantic==2.1.1
pysnmp-lextudio==5.0.28
PyYAML==6.0.1
